Processor
check out the processor. Grinded on this for all of sunday, I thought i would get it done and ordered but I actually want to verify the DRCs on this one so holding off.

This one needs control lines to be soldered by hand as although I have a general idea of the instructions I'm not confident enough to put them in traces.

Should finish this up tmr and order tmr night along with the 74ls245 buffer thingy.